requirements.txt
netifaces
kafka
scapy-python3
numpy
